+++ airlock_processor/StatusChangedQueueTrigger/__init__.py
@@ -6,7 +6,6 @@ import os
 import uuid
 import json
 
-from azure.mgmt.storage import StorageManagementClient
 from exceptions.NoFilesInRequestException import NoFilesInRequestException
 from exceptions.TooManyFilesInRequestException import TooManyFilesInRequestException
 
@@ -23,20 +22,11 @@ class RequestProperties(BaseModel):
 
 class ContainersCopyMetadata:
     source_account_name: str
-    source_account_key: str
-    sa_source_connection_string: str
-    sa_dest_connection_string: str
-    sa_dest_account_name: str
-    sa_dest_resource_group: str
+    dest_account_name: str
 
-    def __init__(self, source_account_name: str, source_account_key: str, sa_source_connection_string: str,
-                 sa_dest_connection_string: str, sa_dest_account_name: str, sa_dest_resource_group: str):
+    def __init__(self, source_account_name: str, dest_account_name: str):
         self.source_account_name = source_account_name
-        self.source_account_key = source_account_key
-        self.sa_source_connection_string = sa_source_connection_string
-        self.sa_dest_connection_string = sa_dest_connection_string
-        self.sa_dest_account_name = sa_dest_account_name
-        self.sa_dest_resource_group = sa_dest_resource_group
+        self.dest_account_name = dest_account_name
 
 
 def main(msg: func.ServiceBusMessage, outputEvent: func.Out[func.EventGridOutputEvent]):
@@ -53,8 +43,6 @@ def main(msg: func.ServiceBusMessage, outputEvent: func.Out[func.EventGridOutput
 
 
 def handle_status_changed(request_properties: RequestProperties):
-    storage_client = blob_operations.get_storage_management_client()
-
     new_status = request_properties.status
     req_id = request_properties.request_id
     ws_id = request_properties.workspace_id
@@ -70,24 +58,20 @@ def handle_status_changed(request_properties: RequestProperties):
 
     if new_status == constants.STAGE_DRAFT and request_type == constants.IMPORT_TYPE:
         account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_EXTERNAL + tre_id
-        account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
-        blob_operations.create_container(account_rg, account_name, req_id, storage_client)
+        blob_operations.create_container(account_name, req_id)
         return
 
     if new_status == constants.STAGE_DRAFT and request_type == constants.EXPORT_TYPE:
         account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INTERNAL + ws_id
-        account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, ws_id)
-        blob_operations.create_container(account_rg, account_name, req_id, storage_client)
+        blob_operations.create_container(account_name, req_id)
         return
 
     if (is_require_data_copy(new_status)):
         logging.info('Request with id %s. requires data copy between storage accounts', req_id)
-        containers_metadata = get_source_dest_env_vars(new_status, request_type, ws_id, storage_client)
-        blob_operations.create_container(containers_metadata.sa_dest_resource_group,
-                                         containers_metadata.sa_dest_account_name, req_id, storage_client)
-        blob_operations.copy_data(containers_metadata.source_account_name, containers_metadata.source_account_key,
-                                  containers_metadata.sa_source_connection_string,
-                                  containers_metadata.sa_dest_connection_string, req_id)
+        containers_metadata = get_source_dest_for_copy(new_status, request_type, ws_id)
+        blob_operations.create_container(containers_metadata.dest_account_name, req_id)
+        blob_operations.copy_data(containers_metadata.source_account_name,
+                                  containers_metadata.dest_account_name, req_id)
         return
 
     # Other statuses which do not require data copy are dismissed as we don't need to do anything...
@@ -117,8 +101,7 @@ def is_require_data_copy(new_status: str):
     return False
 
 
-def get_source_dest_env_vars(new_status: str, request_type: str, short_workspace_id: str,
-                             storage_client: StorageManagementClient) -> ContainersCopyMetadata:
+def get_source_dest_for_copy(new_status: str, request_type: str, short_workspace_id: str) -> ContainersCopyMetadata:
     # sanity
     if is_require_data_copy(new_status) is False:
         raise Exception("Given new status is not supported")
@@ -140,57 +123,30 @@ def get_source_dest_env_vars(new_status: str, request_type: str, short_workspace
         if new_status == constants.STAGE_SUBMITTED:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_EXTERNAL + tre_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_INPROGRESS + tre_id
-            source_account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
-            dest_account_rg = source_account_rg
         elif new_status == constants.STAGE_APPROVAL_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_INPROGRESS + tre_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_APPROVED + short_workspace_id
-            source_account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
-            dest_account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, short_workspace_id)
         elif new_status == constants.STAGE_REJECTION_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_INPROGRESS + tre_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_REJECTED + tre_id
-            source_account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
-            dest_account_rg = source_account_rg
         elif new_status == constants.STAGE_BLOCKING_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_INPROGRESS + tre_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_IMPORT_BLOCKED + tre_id
-            source_account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
-            dest_account_rg = source_account_rg
     else:
         if new_status == constants.STAGE_SUBMITTED:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INTERNAL + short_workspace_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INPROGRESS + short_workspace_id
-            source_account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, short_workspace_id)
-            dest_account_rg = source_account_rg
         elif new_status == constants.STAGE_APPROVAL_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INPROGRESS + short_workspace_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_APPROVED + tre_id
-            source_account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, short_workspace_id)
-            dest_account_rg = constants.CORE_RESOURCE_GROUP_NAME.format(tre_id)
         elif new_status == constants.STAGE_REJECTION_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INPROGRESS + short_workspace_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_REJECTED + short_workspace_id
-            source_account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, short_workspace_id)
-            dest_account_rg = source_account_rg
         elif new_status == constants.STAGE_BLOCKING_INPROGRESS:
             source_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_INPROGRESS + short_workspace_id
             dest_account_name = constants.STORAGE_ACCOUNT_NAME_EXPORT_BLOCKED + short_workspace_id
-            source_account_rg = constants.WORKSPACE_RESOURCE_GROUP_NAME.format(tre_id, short_workspace_id)
-            dest_account_rg = source_account_rg
 
-    logging.info("source [account: '%s', rg: '%s']. dest [account: '%s', rg: '%s']", source_account_name,
-                 source_account_rg, dest_account_name, dest_account_rg)
-
-    sa_source = blob_operations.get_storage_connection_string(source_account_name, source_account_rg, storage_client)
-    sa_dest = blob_operations.get_storage_connection_string(dest_account_name, dest_account_rg, storage_client)
-
-    return ContainersCopyMetadata(sa_source.account_name,
-                                  sa_source.account_key,
-                                  sa_source.connection_string,
-                                  sa_dest.connection_string,
-                                  sa_dest.account_name,
-                                  dest_account_rg)
+    return ContainersCopyMetadata(source_account_name, dest_account_name)
 
 
 def report_failure(outputEvent, request_properties, failed_reason):