+++ templates/core/terraform/airlock/variables.tf
@@ -4,6 +4,7 @@ variable "resource_group_name" {}
 variable "airlock_storage_subnet_id" {}
 variable "airlock_events_subnet_id" {}
 variable "enable_local_debugging" {}
+variable "myip" {}
 variable "api_principal_id" {}
 
 variable "docker_registry_server" {