+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **BREAKING CHANGES & MIGRATIONS**:
 
 * Guacamole workspace service configures firewall requirements with deployment pipeline ([#2371](https://github.com/microsoft/AzureTRE/pull/2371)). **Migration** is manual - update the templateVersion of `tre-shared-service-firewall` in Cosmos to `0.4.0` in order to use this capability.
+* Workspace now has an AirlockManager role that has the permissions to review airlock requests  ([#2349](https://github.com/microsoft/AzureTRE/pull/2349)).
 
 FEATURES:
 
@@ -16,7 +17,7 @@ ENHANCEMENTS:
 
 BUG FIXES:
 
-*
+* Airlock processor creates SAS tokens with _user delegated key_ ([#2382](https://github.com/microsoft/AzureTRE/pull/2376))
 
 ## 0.4.0 (July 27, 2022)